Added tokenIssue() method to keystone API.
This method handles sending an authorization request against the
keystone API, and returns the issued Authorization Token. At this
time it only supports the password authorization plugin. Additional
use cases should be provided by third-party developers.

It is the long term intent that this method will deprecate the
authorize() method, in order to provide a consistently idempotent
implementation.

Change-Id: If1368bcdae251606d2e03043345cd4fc6f0b8b80